requirements.txt
torch
torchvision
numpy
Pillow
opencv-python
rembg